feat: implement condition when user authenticated save & delete from session

+++ order/views.py
@@ -7,14 +7,16 @@ from django.views import View
 class AddProductCookieView(View):
 
     def post(self, request):
-        # if user authenticated save to session else save to cookie
         post_body = json.loads(request.body.decode('utf-8'))
         product_id = post_body.get('product', None)
+        quantity = request.POST.get('quantity', 1)  # this line maybe have problem
         res = HttpResponse('')
         if product_id:
-            quantity = request.POST.get('quantity', 1)
-            res.set_cookie(f'p{product_id}', str(quantity))
-        return res
+            if request.user.is_authenticated:
+                request.session[f'p{product_id}'] = str(quantity)
+            else:
+                res.set_cookie(f'p{product_id}', str(quantity))
+            return res
 
 
 class DeleteProductFromCookieView(View):
@@ -23,7 +25,9 @@ class DeleteProductFromCookieView(View):
         request_body = json.loads(request.body.decode('utf-8'))
         product_id = request_body.get('product', None)
         res = HttpResponse('')
-        print(product_id)
         if product_id:
-            res.delete_cookie(f'p{product_id}')
+            if request.user.is_authenticated:
+                del request.session[f'p{product_id}']
+            else:
+                res.delete_cookie(f'p{product_id}')
         return res
